+++ train_models.py
@@ -32,7 +32,7 @@ parser.add_argument('--workers', type=int, default=16, help='number of data load
 parser.add_argument('--pretrained', type=str, default='', help='path to pretrained model')
 parser.add_argument('--gpu_ids', type=str, default='0,1,2,3', help='comma-seperated string of gpu ids to use for acceleration (-1 for cpu only)')
 parser.add_argument('--input_size', type=int, default=-1, help='input size of network; use -1 to use default input size')
-parser.add_argument('--inc_contrast', type=int, default=1, help='factor to increase contrast for images')
+parser.add_argument('--inc_contrast', type=float, default=1, help='factor to increase contrast for images')
 # Hyperparameters
 parser.add_argument('--epochs', type=int, default=100, help='number of epochs to train for')
 parser.add_argument('--optimizer', type=str, default='sgd', help='optimizer to use')
